backend_api/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
fastapi==0.129.0
greenlet==3.3.1
h11==0.16.0
idna==3.11
joblib==1.5.3
numpy==2.4.2
pandas==3.0.0
psycopg2==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.52.1
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.40.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
APScheduler==3.10.4
